third_party/blink/web_tests/platform/mac/wpt_internal/compute-pressure/losing-focus-suspends-data-delivery.https.window-expected.txt

This is a testharness.js-based test.
FAIL Observer should not receive PressureRecord if page loses focus assert_equals: expected "fair" but got "nominal"
Harness: the test ran to completion.
